% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rcp.R
\name{get_rcp_ma}
\alias{get_rcp_ma}
\title{Return realclearpolitics moving average based on the id of the race}
\usage{
get_rcp_ma(id = NULL, ...)
}
\arguments{
\item{id}{The RCP id of the race}

\item{...}{Additional arguments}
}
\value{
A tibble with the question, answer, date, and percent
}
\description{
Return realclearpolitics moving average based on the id of the race
}
\examples{
# Time series RCP moving average for the 2018 senate race between Sherrod Brown and Jim Renacci.
d <- get_rcp_ma(id = 6331)
}
