% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rcp.R
\name{tidy.rcp_table}
\alias{tidy.rcp_table}
\title{Tidy rcp table}
\usage{
\method{tidy}{rcp_table}(x, ...)
}
\arguments{
\item{x}{an object of class rcp_table}

\item{...}{additional arguments.}
}
\description{
Tidy rcp table
}
